# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Gestor de Estoque API",
    "description": "API REST para gestão de empresas, usuários, produtos, combos, promoções e movimentações de estoque.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Ambiente local" }
  ],
...
  "paths": {
...
    "/api/usuarios": {
      "get": {
        "tags": ["Usuários"],
        "summary": "Listar usuários da própria empresa",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de usuários da empresa do usuário autenticado",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Usuario" } } } }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      },
      "post": {
        "tags": ["Usuários"],
        "summary": "Criar usuário (cadastro público, self-service)",
        "description": "Cada empresa só pode ter um usuário (administrador). Retorna 409 se a empresa informada já tiver um cadastrado.",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UsuarioCreate" } } }
        },
        "responses": {
          "201": {
            "description": "Usuário criado",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Usuario" } } }
          },
          "409": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token obtido em POST /api/auth/login. Envie como 'Authorization: Bearer <token>'."
      }
    },
...
    "responses": {
...
      "Error": {
        "description": "Erro",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": { "message": { "type": "string" } }
      },
...
      "EmpresaResumo": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "nome": { "type": "string" },
          "slug": { "type": "string" }
        }
      },
...
      "Usuario": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "nome": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["ADMIN"] },
          "ativo": { "type": "boolean" },
          "empresaId": { "type": "string", "format": "uuid" },
          "empresa": { "$ref": "#/components/schemas/EmpresaResumo" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "UsuarioCreate": {
        "type": "object",
        "required": ["nome", "email", "senha", "empresaId"],
        "properties": {
          "nome": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "senha": { "type": "string", "format": "password" },
          "empresaId": { "type": "string", "format": "uuid" }
        }
      },
...
    }
  }
}